backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.1
httpx==0.28.1
pydantic==2.9.2
python-dotenv==1.0.1
email-validator==2.2.0